Reduce index size to ushort

+++ inc/obj.hpp
@@ -10,8 +10,8 @@ struct ObjHeader
     ubyte objHeaderSize = sizeof(ObjHeader);
     ubyte shtEntrySize = sizeof(SectionEntry);
     uint shtOffset; // section header table offset in obj file
-    uint shtSize; // section header table size (in entries)
-    uint strEntryId; // entry of names section in section header table
+    ushort shtSize; // section header table size (in entries)
+    ushort strEntryId; // entry of names section in section header table
 };
 
 #endif
